Item Mailboxes
For simple cross-planet transfers. Players can setup the mailbox once obtained (more on that later) and send the contents of the container (self) to the recipients mailbox, if initialized and setup.

So if I am in house near the Janta POI and I want to go send my loots to my mailbox in my city in Norg Bralsin I could.

Note, mailboxes are one per character so no one can have more than one.

Signed-off-by: Bubba Joe <89060381+BubbaJoeX@users.noreply.github.com>